Backend/llms/groq_client.py
import os
from langchain_groq import ChatGroq
from dotenv import load_dotenv

load_dotenv()

llm = ChatGroq(
    model=os.getenv("GROQ_MODEL", "llama-3.3-70b-versatile"),
    temperature=0.3,
    max_tokens=800,
    api_key=os.getenv("GROQ_API_KEY")
)