Backend/requirement.txt
fastapi
uvicorn
langchain
langchain-groq
chromadb
sentence-transformers
requests
beautifulsoup4
pdfplumber
python-dotenv
python-multipart
langchain-core